Switch to CMake to build the shared library. (#3)
This allows us to link to HDF5 in a more controlled manner. We also need to
revamp the wheel builds so that they fetch the correct prebuilt binaries for
HDF5 (and all transitive dependencies). Also, we no longer need assorthead, the
header fetchers, or the manifest file for the headers.

I also tried to build on Windows, but it seems like the wheel delocation tool
is not yet good enough, so we'll just comment that out for now.