Don't bother popping individual CSV columns to free memory.
This fixes a bug where we attempted to re-use the column after popping.
It shouldn't really impact memory efficiency as the C++ and Python heaps
are probably different anyway, so it's freeing the former might not
actually provide more space for the latter.

Nonetheless, we still delete the handle explicitly at function end to
release memory for use by C++ code in other functions.