Support custom validator registration for use with validate_object. (#22)
This differs from alabaster.base's mechanism in that it does not save functions
on the C++ side, to avoid problems with the GIL upon destruction of globals.
Instead, we use a Python-side registry and we create a new takane::Options on
every call into validate_object(). A bit less efficient but it avoids a crash
and is more consistent with the read_object registry anyway.